requirements.txt
Faker==37.0.0
greenlet==3.1.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
SQLAlchemy==2.0.39
typing_extensions==4.12.2
tzdata==2025.1
